+++ changelog.rst
@@ -4,3 +4,4 @@ CHANGELOG
 0.1.0
 -----
 - Add stub read, eval, print functions
+- Add test generator from MAL text tests